requirements.txt
discord.py==1.0.1
tzlocal==1.5.1